Capture mDNS task handle
